requirements.txt
Flask>=2.1.1
Flask-Caching>=1.10.1
Pillow>=9.1.0
PyYAML>=6.0
redis>=4.2.1
uWSGI
